+++ services/BookBuilder/markdown.ts
@@ -1,5 +1,3 @@
-import { MarkdownTree } from './markdownTree'
-
 export class Markdown {
   content: string
 
@@ -13,19 +11,5 @@ export class Markdown {
 
   updateContent = (newContent: string) => {
     this.content = newContent
-  }
-}
-
-export abstract class MarkdownTreeMarkdownAdapter {
-  public static markdownTreeToMarkdown = (tree: MarkdownTree): Markdown => {
-    const markdown = new Markdown('')
-
-    tree.sections.forEach((section) => {
-      section.subsections.forEach((subsection) => {
-        markdown.addText(subsection.content.content)
-      })
-    })
-
-    return markdown
   }
 }
